pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marzlive-upgrade"
version = "0.1.0"
description = "Modern rewrite of Marzlive social platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "tornado>=6.4,<7",
  "motor>=3.6",
  "pymongo>=4.9",
  "pydantic>=2.9",
  "email-validator>=2.2",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "google-cloud-storage>=2.18",
  "google-auth>=2.35",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
  "ruff>=0.6",
  "mypy>=1.11",
]

[tool.setuptools]
packages = ["app"]
